[cat] adding a "hello world" as a proof-of-concept
This is the first utility to experiment with this concept. `cat` will be
like UNIX's `cat`, but it will parse and "reformat" the file to the
output.

The parser leverages my own STL library. I reckon it adds a massive
weight to the compilation of this project, but eventually that library
should allow partial compilation - and the benefits of its parsing
framework are too good to be ignored. And that's why I have that library
as well... :)